Fix uint64 to uint

+++ domains/models/artist.go
@@ -5,7 +5,7 @@ import "github.com/jinzhu/gorm"
 type Artist struct {
 	gorm.Model
 	Album     *Album `gorm:"foreignkey:AlbumID";json:album`
-	AlbumID   uint64 `json:album_id`
+	AlbumID   uint   `json:album_id`
 	Name      string `json:title`
 	Biography string `json:biography`
 }
